Move git config to config/scm
