+++ src/main/groovy/com/cloudogu/gitops/config/MultiTenantSchema.groovy
@@ -1,8 +1,8 @@
 package com.cloudogu.gitops.config
 
-import com.cloudogu.gitops.features.git.config.ScmCentralSchema.GitlabCentralConfig
-import com.cloudogu.gitops.features.git.config.ScmCentralSchema.ScmManagerCentralConfig
-import com.cloudogu.gitops.features.git.config.util.ScmProviderType
+import com.cloudogu.gitops.config.scm.ScmCentralSchema.GitlabCentralConfig
+import com.cloudogu.gitops.config.scm.ScmCentralSchema.ScmManagerCentralConfig
+import com.cloudogu.gitops.config.scm.util.ScmProviderType
 
 import com.fasterxml.jackson.annotation.JsonPropertyDescription
 import picocli.CommandLine.Mixin
