+++ src/main/groovy/com/cloudogu/gitops/config/scm/ScmCentralSchema.groovy
@@ -1,9 +1,9 @@
-package com.cloudogu.gitops.features.git.config
+package com.cloudogu.gitops.config.scm
 
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.config.Credentials
-import com.cloudogu.gitops.features.git.config.util.GitlabConfig
-import com.cloudogu.gitops.features.git.config.util.ScmManagerConfig
+import com.cloudogu.gitops.config.scm.util.GitlabConfig
+import com.cloudogu.gitops.config.scm.util.ScmManagerConfig
 
 import com.fasterxml.jackson.annotation.JsonPropertyDescription
 import picocli.CommandLine.Option
