+++ src/main/groovy/com/cloudogu/gitops/config/scm/util/ScmProviderType.groovy
@@ -1,0 +1,6 @@
+package com.cloudogu.gitops.config.scm.util
+
+enum ScmProviderType {
+	GITLAB,
+	SCM_MANAGER
+}
