--- src/main/groovy/com/cloudogu/gitops/features/git/config/util/ScmProviderType.groovy
@@ -1,6 +1,0 @@
-package com.cloudogu.gitops.features.git.config.util
-
-enum ScmProviderType {
-	GITLAB,
-	SCM_MANAGER
-}
