+++ src/test/groovy/com/cloudogu/gitops/ApplicationTest.groovy
@@ -1,12 +1,13 @@
 package com.cloudogu.gitops
 
-import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.git.config.ScmTenantSchema
-import groovy.transform.CompileStatic
-import io.micronaut.context.ApplicationContext
-import org.junit.jupiter.api.Test
-
 import static org.assertj.core.api.Assertions.assertThat
+
+import com.cloudogu.gitops.config.Config
+import com.cloudogu.gitops.config.scm.ScmTenantSchema
+
+import io.micronaut.context.ApplicationContext
+
+import org.junit.jupiter.api.Test
 
 class ApplicationTest {
