+++ src/test/groovy/com/cloudogu/gitops/git/providers/scmmanager/ScmManagerUrlResolverTest.groovy
@@ -1,20 +1,20 @@
 package com.cloudogu.gitops.git.providers.scmmanager
 
+import static org.junit.jupiter.api.Assertions.*
+import static org.mockito.ArgumentMatchers.any
+import static org.mockito.ArgumentMatchers.eq
+import static org.mockito.Mockito.*
+
+import com.cloudogu.gitops.config.Config
+import com.cloudogu.gitops.config.scm.ScmTenantSchema
+import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.utils.NetworkingUtils
+
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
-
-import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.git.config.ScmTenantSchema
-import com.cloudogu.gitops.kubernetes.api.K8sClient
-import com.cloudogu.gitops.utils.NetworkingUtils
-
-import static org.junit.jupiter.api.Assertions.*
-import static org.mockito.ArgumentMatchers.any
-import static org.mockito.ArgumentMatchers.eq
-import static org.mockito.Mockito.*
 
 @ExtendWith(MockitoExtension.class)
 class ScmManagerUrlResolverTest {
